+++ runtime/src/main/java/com/instana/android/performance/PerformanceMetric.kt
@@ -10,4 +10,5 @@ internal sealed class PerformanceMetric {
     data class AppStartTime(val coldStart: Long = 0L,val warmStart:Long= 0L, val hotStart:Long= 0L) : PerformanceMetric()
     data class AppNotResponding(val duration: Long? = 0L,val stackTrace:String,val allStackTrace:String) : PerformanceMetric()
     data class OutOfMemory(val availableMb:Long, val usedMb:Long, val maximumMb:Long) : PerformanceMetric()
+    data class ExcessiveBackgroundNetworkUsage(val usedMb:Long) : PerformanceMetric()
 }