+++ runtime/src/test/java/com/instana/android/core/InstanaWorkManagerTest.kt
@@ -148,7 +148,6 @@ class InstanaWorkManagerTest : BaseTest() {
         val queue: Queue<Beacon> = LinkedList()
         queue.offer(beacon)
         VisibleScreenNameTracker.initialViewMap = mapOf(
-            ScreenAttributes.ACTIVITY_RESUME_TIME.value to System.nanoTime().toString(),
             ScreenAttributes.ACTIVITY_CLASS_NAME.value to "Test1",
             ScreenAttributes.ACTIVITY_LOCAL_PATH_NAME.value to "Test2",
             ScreenAttributes.ACTIVITY_SCREEN_NAME.value to "Test3"
@@ -190,7 +189,6 @@ class InstanaWorkManagerTest : BaseTest() {
         queue.offer(beacon)
         beacon.setInternalMeta(ScreenAttributes.ACTIVITY_SCREEN_NAME.value,"Test3")
         VisibleScreenNameTracker.initialViewMap = mapOf(
-            ScreenAttributes.ACTIVITY_RESUME_TIME.value to System.nanoTime().toString(),
             ScreenAttributes.ACTIVITY_CLASS_NAME.value to "Test1",
             ScreenAttributes.ACTIVITY_LOCAL_PATH_NAME.value to "Test2",
             ScreenAttributes.ACTIVITY_SCREEN_NAME.value to "Test3"